Remove govuk-modules directory from gitignore
This directory is required for SASS compilation, but was in the gitignore file, with no apparant script to regenerate it. It looks to be a local copy or a modification to the previous govuk frontend to prevent it being changed on new releases.

Having this 'fixed' version of the front-end kit locally should resolve issues with SASS compilation